MET-6370 Review remarks
